Cambios de velocidad

+++ snake/matriz.cpp
@@ -49,8 +49,8 @@ int playmatrix(void){
     
     //int flag =0;
     thread th1(keyEventSnake , &snake);
-    int timer = 0;
-    int speed = 1;
+    double timer = 0;
+    double speed = 1;
     while (snake.flag)
     {
         if(snake.redirect || timer == 10*speed){
